+++ test/bigtest/interactors/transactions/TransactionDetailsInteractor.js
@@ -1,4 +1,5 @@
 import {
+  clickable,
   interactor,
   isPresent,
   text,
@@ -10,11 +11,17 @@ import Button from '../common/Button';
   value = text('[data-test-transaction-status-value]');
 }
 
+@interactor class SourceLink {
+  static defaultScope = '[data-testid="transaction-source-link"]';
+  click = clickable();
+}
+
 export default interactor(class TransactionDetailsInteractor {
   static defaultScope = '#pane-transaction-details';
 
   closePane = new Button('[icon=times]');
   status = new Status();
+  sourceLink = new SourceLink();
 
   isLoaded = isPresent('[class*=paneTitleLabel---]');
   whenLoaded() {